+++ src/modules/search/components/ClientSearch.tsx
@@ -10,7 +10,8 @@ import ClientDisplayTypeToggle, {
   DisplayType,
 } from './ClientDisplayTypeToggle';
 import ClientSearchTypeToggle, { SearchType } from './ClientSearchTypeToggle';
-import { ClientTextSearchInputForm } from './ClientTextSearchInput';
+
+import ClientTextSearchForm from './ClientTextSearchForm';
 import ButtonLink from '@/components/elements/ButtonLink';
 import { externalIdColumn } from '@/components/elements/ExternalIdDisplay';
 import { ColumnDef } from '@/components/elements/table/types';
@@ -256,7 +257,7 @@ const ClientSearch = () => {
       </Stack>
       <Box mb={5}>
         {searchType === 'broad' ? (
-          <ClientTextSearchInputForm
+          <ClientTextSearchForm
             initialValue={initialValues?.textSearch || ''}
             onSearch={(text) => handleSubmitSearch({ textSearch: text })}
             label={null}
